fixed long in em2da_port

+++ petram/phys/em2da/em2da_port.py
@@ -222,7 +222,7 @@ class EM2Da_Port(EM2Da_Bdry):
     def import_panel1_value(self, v):
         self.port_idx = v[0]       
         self.mode = v[1]
-        self.mn = [long(x) for x in v[2].split(',')]
+        self.mn = [int(x) for x in v[2].split(',')]
         self.vt.import_panel_value(self, v[3:])                         
 
     def get_exter_NDoF(self):
